requirements.txt
autogen-agentchat
autogen-core
autogen-ext
asyncio
dotenv
openai
tiktoken
autogen-ext[docker]
streamlit
langgraph
llama-index
langchain-tavily
fastapi
uvicorn
loguru
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-embeddings-google-genai
llama-index-llms-google-genai 
pymupdf
chonkie
llama-index 
llama-index-vector-stores-chroma 
llama-index-vector-stores-faiss 
chromadb
faiss-cpu
tavily